[UPDATE] seperated components
